Create first few functions to handle variable extractions

+++ prog/src/symbols.hpp
@@ -5,7 +5,7 @@
 #include <unordered_map>
 #include <variant>
 
-using AnyType = std::variant<int, float, bool, char, std::string>;
+using AnyType = std::variant<int, double, bool, char, std::string>;
 using Variables = std::unordered_map<std::string, AnyType>;
 
 class SymbolTable
@@ -25,8 +25,8 @@ inline std::ostream& operator<<(std::ostream& os, const AnyType& any)
     if (std::holds_alternative<int>(any)) {
         os << std::get<int>(any) << std::endl;
     }
-    else if (std::holds_alternative<float>(any)) {
-        os << std::get<float>(any) << std::endl;
+    else if (std::holds_alternative<double>(any)) {
+        os << std::get<double>(any) << std::endl;
     }
     else if (std::holds_alternative<char>(any)) {
         os << std::get<char>(any) << std::endl;
